IntelliHealthApp/IntelliHealthFlask/requirements.txt
Flask
flask_cors
joblib==1.3.1
langchain
numpy
openai==0.27.8
pandas
pinecone-client
python-dotenv==1.0.0
scikit_learn
tensorflow
lightgbm